+++ operator/pkg/resources/target/statefulset.go
@@ -92,6 +92,7 @@ func targetPodSpec(ais *aisv1.AIStore) *corev1.PodSpec {
 				Resources:       *cmn.NewInitResourceReq(),
 				Args:            cmn.NewInitContainerArgs(aisapc.Target, ais.Spec.HostnameMap),
 				VolumeMounts:    cmn.NewInitVolumeMounts(),
+				SecurityContext: cmn.RestrictedSecurityContext(),
 			},
 		},
 		Containers: []corev1.Container{